Extract AndroidMvpView
