+++ piano-hero/mobile/src/main/java/com/novoda/pianohero/AndroidThingsActivity.java
@@ -8,10 +8,9 @@ import android.view.View;
 
 import static android.view.View.GONE;
 
-public class AndroidThingsActivity extends AppCompatActivity implements GameMvp.View {
+public class AndroidThingsActivity extends AppCompatActivity {
 
     private GamePresenter gamePresenter;
-    private GameScreen gameScreen;
     private Speaker speaker;
     private TouchButton touchButton;
 
@@ -31,8 +30,8 @@ public class AndroidThingsActivity extends AppCompatActivity implements GameMvp.
         ViewModelConverter viewModelConverter = new ViewModelConverter(simplePitchNotationFormatter);
         SongPlayer songPlayer = new SongPlayer();
         GameModel gameModel = new GameModel(songSequenceFactory, piano, touchButton, viewModelConverter, songPlayer);
-        gameScreen = (GameScreen) findViewById(R.id.game_screen);
-        gamePresenter = new GamePresenter(gameModel, this);
+        GameScreen gameScreen = (GameScreen) findViewById(R.id.game_screen);
+        gamePresenter = new GamePresenter(gameModel, new AndroidGameMvpView(gameScreen, speaker));
 
         gamePresenter.onCreate();
     }
@@ -58,61 +57,6 @@ public class AndroidThingsActivity extends AppCompatActivity implements GameMvp.
     protected void onResume() {
         super.onResume();
         gamePresenter.onResume();
-    }
-
-    @Override
-    public void startSound(double midi) {
-        speaker.start(midi);
-    }
-
-    @Override
-    public void stopSound() {
-        speaker.stop();
-    }
-
-    @Override
-    public void showClock(ClockViewModel viewModel) {
-        gameScreen.showClock(viewModel);
-    }
-
-    @Override
-    public void showGameStarted(GameStartViewModel viewModel) {
-        gameScreen.showGameStart(viewModel);
-    }
-
-    @Override
-    public void showSong(SongStartViewModel viewModel) {
-        gameScreen.showSongStart(viewModel);
-    }
-
-    @Override
-    public void showRound(RoundEndViewModel viewModel) {
-        gameScreen.showSuccess(viewModel);
-    }
-
-    @Override
-    public void showSongComplete(SongCompleteViewModel viewModel) {
-        gameScreen.showSongComplete(viewModel);
-    }
-
-    @Override
-    public void showError(RoundErrorViewModel viewModel) {
-        gameScreen.showError(viewModel);
-    }
-
-    @Override
-    public void showSharpError(RoundErrorViewModel viewModel) {
-        gameScreen.showSharpError(viewModel);
-    }
-
-    @Override
-    public void showScore(String score) {
-        Log.d("!!", "Update " + score);
-    }
-
-    @Override
-    public void showGameComplete(GameOverViewModel viewModel) {
-        gameScreen.showGameComplete(viewModel);
     }
 
     @Override
